+++ Loadout_40K/Weapon/LT_wpn_Guard.sqf
@@ -65,12 +65,4 @@ _handGunAttScope = [""];
 _launcher = ["IC_CAD_Rocketlauncher_base"];
 _launcher_Mags = ["IC_Frag_rocket_mag","IC_Krak_rocket_mag"];
 
-_launcherAA = ["IC_Launcher_AA_grey"];
-_launcherAA_Mags = ["IC_Flak_Missile_mag"];
-
-_heavyAR = ["ICP_LMG_01_imp"];
-_heavyAR_Mags = "ic_100rnd_825_HP_Tracer_rifleMag";
-
-_binocular = ["ic_cad_magnoculars"];
-
 //=====================================================================================
