+++ Loadout_40K/Weapon/LT_wpn_Ork.sqf
@@ -65,12 +65,4 @@ _handGunAttScope = [""];
 _launcher = ["DKOK_Mole_Launcher_1"];
 _launcher_Mags = ["Mol_AT_Mag","Mol_HE_Mag"];
 
-_launcherAA = ["IC_Launcher_AA_grey"];
-_launcherAA_Mags = ["IC_Flak_Missile_mag"];
-
-_heavyAR = ["ICP_LMG_01_imp"];
-_heavyAR_Mags = "ic_100rnd_825_HP_Tracer_rifleMag";
-
-_binocular = [""];
-
 //=====================================================================================
