+++ Loadout_GM/Weapon/LT_wpn_G3A3.sqf
@@ -64,12 +64,4 @@ _handGunAttScope = [""];
 _launcher = ["gm_pzf44_2_oli_feroz2x17"];
 _launcher_Mags = ["gm_1Rnd_44x537mm_heat_dm32_pzf44_2"];
 
-_launcherAA = ["gm_fim43_oli"];
-_launcherAA_Mags = ["gm_1Rnd_70mm_he_m585_fim43","Titan_AA"];
-
-_heavyAR = ["gm_lmgm62_blk"];
-_heavyAR_Mags = "gm_120Rnd_762x51mm_B_T_DM21A2_mg3_grn";
-
-_binocular = ["gm_ferod16_oli"];
-
 //=====================================================================================
