+++ Loadout_GM/Weapon/LT_wpn_NLD_90.sqf
@@ -66,12 +66,4 @@ _handGunAttScope = [""];
 _launcher = ["gm_pzf84_win_feroz2x17"];
 _launcher_Mags = ["gm_1Rnd_84x245mm_heat_t_DM22_carlgustaf"];
 
-_launcherAA = ["gm_fim43_oli"];
-_launcherAA_Mags = ["gm_1Rnd_70mm_he_m585_fim43","Titan_AA"];
-
-_heavyAR = ["gm_lmgm62_blk"];
-_heavyAR_Mags = "gm_120Rnd_762x51mm_B_T_DM21A2_mg3_grn";
-
-_binocular = ["gm_ferod16_oli"];
-
 //=====================================================================================
