+++ Loadout_GM/paramsGM.hpp
@@ -90,15 +90,15 @@ class lt_gear_weapon_green
 class lt_gear_SR
 {
     title = "Short range radio settings";
-    values[] = {0,1,2,3,4,5,6,7};
-    texts[] = {"Nobody","Everyone","Commandant/Squad Leader","Commandant/Squad Leader/Fire Team Leader","Commandant/Squad Leader/Fire Team Leader/Combat Life Safer","Commandant/Squad Leader/Fire Team Leader/Genadier","Commandant/Squad Leader/Fire Team Leader/Dedicated markmen","Commandant/Squad Leader/Fire Team Leader/Radioman(GM/VN)"};
+    values[] = {0,1};
+    texts[] = {"Nobody","Everyone"};
     default = SHORT_RADIO;
 };
 class lt_gear_radiopack
 {
     title = "Long range radio settings";
-    values[] = {0,1};
-    texts[] = {"Nobody","Radioman(GM/VN)"};
+    values[] = {0,1,2};
+    texts[] = {"Nobody","Radioman(GM/VN)","Role Specific"};
     default = LONG_RADIO;
 };
 class lt_gear_nvg
