--- Reference/EmptyTemplates/LT_camo_.sqf
@@ -1,43 +1,0 @@
-/*
-
-File: LT_camo_.sqf
-Tijdperk: GM
-Author: Brainless_Ben
-
-Description:
-Reset player weapons set from mission parameters
-
-*/
-
-//=====================================================================================
-
-// Define clothing that will be used
-_unif = [""];
-_vest = [""];
-_helm = [""];
-	//Optioneel medic kleding
-_unifMedic = [""];
-_vestMedic = [""];
-_helmMedic = [""];
-
-_nvgs = [""];
-_goggles = [""];
-
-_unifCrew = [""];
-_vestCrew = [""];
-_helmCrew = [""];
-
-_unifHeliP = [""];
-_unifHeliC = [""];
-_vestHeli = [""];
-_helmHeli = [""];
-
-_unifJet = [""];
-_helmJet = [""];
-
-_backpackSmall = [""];
-_backpackLarge = [""];
-_backpackMedic = [""];
-_backpackRadio = [""];
-
-//=====================================================================================
