+++ script_component.hpp
@@ -1,0 +1,3 @@
+#define COMPONENT template_gear
+#define COMPONENT_BEAUTIFIED Templat_Gear
+#include "\x\cba\addons\main\script_macros_common.hpp"
