--- script_version.hpp
@@ -1,7 +1,0 @@
-#define MAJOR 4
-#define MINOR 55
-#define BUILD 35
-
-#define VERSION MAJOR.MINOR
-#define VERSION_STR MAJOR.MINOR.BUILD
-#define VERSION_AR MAJOR,MINOR,BUILD
